+++ languages/en/default.php
@@ -11,15 +11,14 @@
  * @license LGPL
  */
 
-
 /**
  * Miscellaneous
  */
-$GLOBALS['TL_LANG']['MSC']['eventSubscribe']    = 'Subscribe';
-$GLOBALS['TL_LANG']['MSC']['eventUnsubscribe']  = 'Unsubscribe';
-$GLOBALS['TL_LANG']['MSC']['eventSubscribed']   = 'You have subscribed to the event.';
-$GLOBALS['TL_LANG']['MSC']['eventUnsubscribed'] = 'You have unsubscribed from the event.';
-
+$GLOBALS['TL_LANG']['MSC']['eventSubscribe']              = 'Subscribe';
+$GLOBALS['TL_LANG']['MSC']['eventUnsubscribe']            = 'Unsubscribe';
+$GLOBALS['TL_LANG']['MSC']['eventSubscribed']             = 'You have subscribed to the event.';
+$GLOBALS['TL_LANG']['MSC']['eventUnsubscribed']           = 'You have unsubscribed from the event.';
+$GLOBALS['TL_LANG']['MSC']['eventSubscriptionNotAllowed'] = 'It is no longer possible to subscribe to this event.';
 
 /**
  * Errors