+++ src/Codefog/EventsSubscriptions/EventsSubscriptions.php
@@ -13,10 +13,60 @@
 
 namespace Codefog\EventsSubscriptions;
 
+use Contao\CalendarEventsModel;
 use Contao\Database;
 
 class EventsSubscriptions
 {
+    /**
+     * Return true if the member can subscribe to the event
+     *
+     * @param int $eventId
+     * @param int $memberId
+     *
+     * @return bool
+     */
+    public static function canSubscribe($eventId, $memberId)
+    {
+        if (($event = CalendarEventsModel::findByPk($eventId)) === null) {
+            return false;
+        }
+
+        // Already subscribed to the event
+        if (static::isSubscribed($eventId, $memberId)) {
+            return false;
+        }
+
+        // Validate maximum number of subscriptions
+        if (!static::validateMaximumSubscriptions($event)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * Validate the maximum number of subscriptions
+     *
+     * @param CalendarEventsModel $event
+     *
+     * @return bool
+     */
+    private static function validateMaximumSubscriptions(CalendarEventsModel $event)
+    {
+        if (!$event->subscription_maximum) {
+            return true;
+        }
+
+        $total = Database::getInstance()->prepare(
+            "SELECT COUNT(*) AS total FROM tl_calendar_events_subscriptions WHERE pid=?"
+        )
+            ->execute($event->id)
+            ->total;
+
+        return $total < $event->subscription_maximum;
+    }
+
     /**
      * Subscribe the member and return true on success, false otherwise
      *
@@ -27,7 +77,7 @@ class EventsSubscriptions
      */
     public static function subscribeMember($eventId, $memberId)
     {
-        if (static::checkSubscription($memberId, $memberId)) {
+        if (!static::isSubscribed($memberId, $memberId)) {
             $insertId = Database::getInstance()->prepare(
                 "INSERT INTO tl_calendar_events_subscriptions (tstamp, pid, member) VALUES (?, ?, ?)"
             )
@@ -52,7 +102,7 @@ class EventsSubscriptions
      */
     public static function unsubscribeMember($eventId, $memberId)
     {
-        if (!static::checkSubscription($eventId, $memberId)) {
+        if (static::isSubscribed($eventId, $memberId)) {
             $affectedRows = Database::getInstance()->prepare(
                 "DELETE FROM tl_calendar_events_subscriptions WHERE pid=? AND member=?"
             )
@@ -68,14 +118,14 @@ class EventsSubscriptions
     }
 
     /**
-     * Return false if the member is already subscribed, true otherwise
+     * Return true if the member is subscribed
      *
      * @param int $eventId
      * @param int $memberId
      *
      * @return boolean
      */
-    public static function checkSubscription($eventId, $memberId)
+    public static function isSubscribed($eventId, $memberId)
     {
         $subscription = Database::getInstance()->prepare(
             "SELECT id FROM tl_calendar_events_subscriptions WHERE pid=? AND member=?"
@@ -83,6 +133,6 @@ class EventsSubscriptions
             ->limit(1)
             ->execute($eventId, $memberId);
 
-        return $subscription->numRows ? false : true;
+        return $subscription->numRows ? true : false;
     }
 }